backend/requirements.txt
absl-py==2.1.0
accessible-pygments==0.0.4
alabaster==0.7.16
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astroid==2.15.8
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
blis==0.7.11
bottle==0.12.25
bson==0.5.10
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
clr-loader==0.2.6
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.2
dnspython==2.6.1
docutils==0.18.1
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
filelock==3.13.4
flasgger==0.9.7.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SSE==1.0.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
furo==2024.1.29
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
imagesize==1.4.1
ipython==8.25.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jira==3.8.0
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
keras==3.2.0
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language-tool-python==2.7.3
language_data==1.2.0
lazy-object-proxy==1.10.0
libclang==18.1.1
lxml==5.1.0
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.7
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.3
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
parso==0.8.4
pdfminer==20191125
pdfminer.six==20231228
pefile==2023.2.7
piccolo-theme==0.19.0
pillow==10.3.0
preshed==3.0.9
prompt_toolkit==3.0.46
protobuf==4.25.3
ProWritingAidSDK @ git+https://github.com/prowriting/prowritingaid.python.git@fd6fc238778b7b5ba4bb56835d3ed73594f3bcb7
proxy_tools==0.1.0
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydata-sphinx-theme==0.15.2
pydot==2.0.0
Pygments==2.17.2
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
pymongo==4.7.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.2
pythonnet==3.0.3
pywebview==5.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
redis==5.0.5
referencing==0.35.1
regex==2023.12.25
renku-sphinx-theme==0.4.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-detector==1.2.2
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
selenium==4.20.0
semver==3.0.2
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==6.1.3
sphinx-basic-ng==1.0.0b2
sphinx-copybutton==0.5.2
sphinx-press-theme==0.8.0
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-email==0.3.6
sphinxcontrib-googleanalytics==0.4
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel
termcolor==2.4.0
tf_keras==2.16.0
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.39.3
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
